体重を更新するアクションを修正
